app/requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
Werkzeug==3.0.3
pytest
pytest-flask
pytest-cov
requests